[glib] use CLOCK_BOOTTIME for monotonic time if requested.
CLOCK_MONOTONIC has the drawback that it does not increment during
suspend. Some applications may need timeouts that follow wall clock
time even if the system is suspended; provide a way to do this by
adding an environment variable G_CLOCK that determines which clock
is used for monotonic time.